Adding functionalities for seamless Windows/Linux build and run

+++ SOURCE_FILES/00_MAIN.cu
@@ -18,13 +18,14 @@
 
 #include "config.hpp"
 #include "Case_definition.cuh"
+#include <string>
 
-int PDF_EVOLUTION(cudaDeviceProp* prop);
+int16_t PDF_EVOLUTION(cudaDeviceProp* prop);
 
-void Intro_square_filler(std::string& message, const u_int16_t border_length, const u_int16_t number_of_tabs, const u_int16_t offset){
+void Intro_square_filler(std::string& message, const uint16_t border_length, const uint16_t number_of_tabs, const uint16_t offset){
 
-	u_int16_t msg_length = message.size() + number_of_tabs*8 - 1 - offset;
-	for (u_int16_t k = 0; k < border_length - msg_length; k++){
+	uint16_t msg_length = message.size() + number_of_tabs*8 - 1 - offset;
+	for (uint16_t k = 0; k < border_length - msg_length; k++){
 		message.append(" ");
 	}
 	message.append("|\n");
@@ -47,7 +48,7 @@ int main() {
 		std::string border_mid 	= "|==========================================================================================|";
 		std::cout << border <<"\n";
 		std::cout << border_mid <<"\n";
-		const u_int16_t border_length = border.size();
+		const uint16_t border_length = border.size();
 
 		std::string temp = "|	Wecome to the Liouville Eq. simulator. You are using version number "; temp.append(project_version);
 		Intro_square_filler(temp, border_length, 1, 0);
